Refactor timer logic into it's component
